Build an utility to loop through the children of a container taking into account the rtl comments
